feat: useCoopShopCafeteria 훅 생성
