+++ src/api/coopshop/APIDetail.ts
@@ -1,15 +1,14 @@
 import { type APIRequest, HTTP_METHOD } from 'interfaces/APIRequest';
-import {
-  CoopshopCafeteriaResponse,
-} from './entity';
+import { CoopshopResponse } from './entity';
 
-export class CoopshopCafeteria<R extends CoopshopCafeteriaResponse> implements APIRequest<R> {
+// /coopshop/1: 학생식당
+// /coopshop/2: 세탁소
+// /coopshop/3: 참빛관 편의점
+
+export class CoopshopCafeteria<R extends CoopshopResponse> implements APIRequest<R> {
   method = HTTP_METHOD.GET;
 
   path = '/coopshop/1';
-  // 1: 학생식당
-  // 2: 세탁소
-  // 3: 참빛관 편의점
 
   response!: R;
