+++ src/api/coopshop/index.ts
@@ -1,4 +1,4 @@
 import APIClient from 'utils/ts/apiClient';
 import { CoopshopCafeteria } from './APIDetail';
 
-export default APIClient.of(CoopshopCafeteria);
+export const getCafeteriaInfo = APIClient.of(CoopshopCafeteria);
